refactor: extract runtime policy contracts
